Move DateTimeUtil.java into a util package
DateTimeUtil was moved into a util package. Code was also checked to ensure it complies with checkstyle.
