+++ src/main/java/haru/storage/Storage.java
@@ -10,13 +10,13 @@ import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 import java.util.Scanner;
 
-import haru.DateTimeUtil;
 import haru.HaruException;
 import haru.task.Deadline;
 import haru.task.Event;
 import haru.task.Task;
 import haru.task.TaskList;
 import haru.task.Todo;
+import haru.util.DateTimeUtil;
 
 /**
  * Performs file operations on the task file.
